ml-service/requirements.txt
# ============================================================
# GigPay ML Service — Python Dependencies (Slimmed for Hackathon)
# ============================================================

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.31.0

# Data validation
pydantic>=2.9.2

# ML model loading & inference
joblib>=1.3.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# File upload support (required by FastAPI UploadFile)
python-multipart>=0.0.9

# Environment
python-dotenv>=1.0.1
